+++ go.mod
@@ -2,4 +2,4 @@ module github.com/AntonyCarl/OMA-Library
 
 go 1.21.6
 
-require github.com/lib/pq v1.10.9 // indirect
+require github.com/lib/pq v1.10.9
